+++ NEWS.md
@@ -1,4 +1,18 @@
+# fable 0.4.0
+
+## New features
+
+* Added `generate()` and `IRF()` methods for VAR models.
+* Added `IRF()` method for ARIMA models.
+* Added `VECM()` and `VARIMA()` models.
+
+## Improvements
+
+* Documentation improvements.
+
 # fable 0.3.4
+
+Small patch to resolve issues in C++ R headers.
 
 ## Improvements